backend/requirements.txt
# FastAPI & server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth & utils
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.17
pydantic-settings==2.6.1

# PDF (문서 다운로드)
weasyprint>=62.0
